requirements.txt
streamlit>=1.30.0
supabase>=2.0.0
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
